refactor: extract applyBuff to shared statusEffectUtils
Move the duplicated "apply or refresh status effect" pattern to a
shared utility. Replaces inline implementations in buffEffectHandler,
zoneEffectHandler, and ServerZoneSystem with a single applyBuff call.

Closes #256
