+++ server/src/game/ServerZoneSystem.ts
@@ -2,8 +2,8 @@ import type { MapSchema } from '@colyseus/schema'
 import type { HeroSchema } from '../schema/HeroSchema.js'
 import type { ZoneSchema } from '../schema/ZoneSchema.js'
 import type { MinionSchema } from '../schema/MinionSchema.js'
-import { StatusEffectSchema } from '../schema/StatusEffectSchema.js'
 import { distanceSq } from '@shared/math/distanceSq'
+import { applyBuff } from './statusEffectUtils.js'
 import { ZONE_EFFECT_DURATION } from '@shared/constants'
 
 function shouldAffect(zone: ZoneSchema, hero: HeroSchema): boolean {
@@ -81,20 +81,8 @@ export function tickZones(
 
       // Apply or refresh status effect (skip for expired zones)
       if (zone.remainingDuration > 0) {
-        const effectKey = zoneId
         const effectDuration = zone.effectDuration > 0 ? zone.effectDuration : ZONE_EFFECT_DURATION
-        const existing = hero.statusEffects.get(effectKey)
-        if (existing) {
-          existing.remainingDuration = effectDuration
-        } else {
-          const effect = new StatusEffectSchema()
-          effect.id = effectKey
-          effect.buffType = zone.buffType
-          effect.value = zone.value
-          effect.remainingDuration = effectDuration
-          effect.isDebuff = zone.isDebuff
-          hero.statusEffects.set(effectKey, effect)
-        }
+        applyBuff(hero, zoneId, zone.buffType, zone.value, effectDuration, zone.isDebuff)
       }
     })
